fix(RPC): version `RPC` same as` rest`, export again in `shortcuts`
