+++ .gitignore
@@ -31,6 +31,11 @@ shortcuts/**/*.map
 shortcuts/**/*.d.ts
 shortcuts/**/*.mjs
 
+utils/**/*.js
+utils/**/*.map
+utils/**/*.d.ts
+utils/**/*.mjs
+
 voice/**/*.js
 voice/**/*.map
 voice/**/*.d.ts
