+++ deno/shortcuts/v8.ts
@@ -2,4 +2,5 @@ export * from '../globals.ts';
 export * from '../gateway/v8.ts';
 export * from '../payloads/v8/mod.ts';
 export * from '../rest/v8/mod.ts';
+export * from '../rpc/v8.ts';
 export * as Utils from '../utils/v8.ts';
