+++ shortcuts/v8.ts
@@ -2,4 +2,5 @@ export * from '../globals';
 export * from '../gateway/v8';
 export * from '../payloads/v8/index';
 export * from '../rest/v8/index';
+export * from '../rpc/v8';
 export * as Utils from '../utils/v8';
